Pass list to the form
